Move Box and Flex to their own respective folders
Change-type: patch
