+++ package.json
@@ -36,8 +36,8 @@
     "test:local-visual": "./scripts/test-screenshots.sh",
     "test:visual": "npm run ci:screenshot -- -o .reg/current && reg-cli .reg/current __screenshots__ .reg/diff -R .reg/report.html",
     "jest": "jest --maxWorkers=4",
-    "eslint": "eslint src **/*.spec.js .storybook",
-    "eslint-fix": "eslint src **/*.spec.js .storybook --fix",
+    "eslint": "eslint src .storybook",
+    "eslint-fix": "eslint src .storybook --fix",
     "tslint": "tslint -t stylish \"src/**/*.ts\" \"src/**/*.tsx\" \"typings/**/*.d.ts\"",
     "tslint-fix": "tslint -t stylish --fix \"src/**/*.ts\" \"src/**/*.tsx\" \"typings/**/*.d.ts\"",
     "lint": "npm run eslint && npm run tslint",
