+++ src/components/DataTypes/object.tsx
@@ -8,7 +8,7 @@ import mapValues from 'lodash/mapValues';
 import * as React from 'react';
 import * as utils from '../../utils';
 import { DataTypeEditProps } from '../Filters';
-import { Flex } from '../Grid';
+import { Flex } from '../Flex';
 import Input from '../Input';
 import { getJsonDescription } from './utils';
