+++ src/components/Filters/Summary.tsx
@@ -2,8 +2,9 @@ import { JSONSchema6 } from 'json-schema';
 import * as React from 'react';
 import FaChartPie from 'react-icons/lib/fa/pie-chart';
 import { FiltersView, ViewScope } from '.';
+import { Box } from '../Box';
 import Button from '../Button';
-import { Box, Flex } from '../Grid';
+import { Flex } from '../Flex';
 import Input from '../Input';
 import Modal from '../Modal';
 import Select from '../Select';
