+++ src/components/Filters/index.tsx
@@ -14,9 +14,10 @@ import styled from 'styled-components';
 import { Button, ButtonProps } from '../../';
 import { DefaultProps } from '../../common-types';
 import * as utils from '../../utils';
+import { Box } from '../Box';
 import { getDataModel } from '../DataTypes';
 import { DropDownButtonProps } from '../DropDownButton';
-import { Box, Flex } from '../Grid';
+import { Flex } from '../Flex';
 import Modal from '../Modal';
 import Search from '../Search';
 import Select from '../Select';
