+++ src/extra/Mermaid/index.tsx
@@ -2,7 +2,7 @@ import { mermaidAPI } from 'mermaid';
 import * as React from 'react';
 import uuid from 'uuid/v4';
 import { DefaultProps } from '../../common-types';
-import { Box, BoxProps } from '../../components/Grid';
+import { Box, BoxProps } from '../../components/Box';
 
 interface MermaidProps extends DefaultProps, BoxProps {
 	value: string;
