+++ src/unstable/components/Form/index.tsx
@@ -5,8 +5,8 @@ import omit from 'lodash/omit';
 import * as React from 'react';
 import Form, { IChangeEvent } from 'react-jsonschema-form';
 import styled from 'styled-components';
+import { Box } from '../../../components/Box';
 import Button from '../../../components/Button';
-import { Box } from '../../../components/Grid';
 import * as utils from '../../../utils';
 import { FormProps } from '../../unstable-typings';
 import { DescriptionField } from './fields/DescriptionField';
